requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
scikit-learn==1.3.1
numpy==1.26.0
pandas==2.1.1
bandit==1.7.5
gitpython==3.1.40
python-jose==3.3.0
python-multipart==0.0.6
matplotlib==3.8.0
seaborn==0.13.0
# tensorflow==2.14.0
transformers==4.37.2
pylint==3.0.2
pytest==7.4.2
python-magic